refactor: moves validator_factory back to trestle.core
The validator_factory imports from trestle.core and create a wrapper
around the validation API for trestle use.
