+++ docs/reference/API/trestle/core/validator_factory.md
@@ -1,0 +1,7 @@
+---
+title: trestle.core.validator_factory
+description: Documentation for trestle.core.validator_factory module
+---
+
+::: trestle.core.validator_factory
+handler: python
