--- docs/reference/API/trestle/sdk/validation/validator_factory.md
@@ -1,7 +1,0 @@
----
-title: trestle.sdk.validation.validator_factory
-description: Documentation for trestle.sdk.validation.validator_factory module
----
-
-::: trestle.sdk.validation.validator_factory
-handler: python
