+++ trestle/sdk/validation/all_validator.py
@@ -17,7 +17,7 @@
 import pathlib
 from typing import Optional
 
-import trestle.sdk.validation.validator_factory as vfact
+import trestle.core.validator_factory as vfact
 from trestle.sdk.models.base_model import OscalBaseModel
 from trestle.sdk.validation.validator import Validator
